lib: fwts_framework: Make string length type a size_t rather than an int
fwts_framework_strtrunc() string length parameter max needs to be a
size_t rather than an int.

Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>

+++ src/lib/src/fwts_framework.c
@@ -287,7 +287,7 @@ static void fwts_framework_show_tests(fwts_framework *fw, bool full)
  *  fwts_framework_strtrunc()
  *	truncate overlong string
  */
-static void fwts_framework_strtrunc(char *dest, const char *src, int max)
+static void fwts_framework_strtrunc(char *dest, const char *src, size_t max)
 {
 	strncpy(dest, src, max);
